refactor move from mod to escrow/token.rs
